+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# Changelog for project fj-doc-mod-itext
+# Changelog for project fj-doc-mod-openpdf
+
+## version 0.1.0 - 2023-08-13
+* Updated parent version to fj-doc 1.4.4
+* Typos in documentation
+* Added JUnit for basic testing
 
 ## version 0.1.0-rc.001 - 2023-07-22
 * Porting of fj-doc-mod-itext based on [OpenPDF](https://github.com/LibrePDF/OpenPDF) / [OpenRTF](https://github.com/LibrePDF/OpenRTF)